[HUBS] Add support to use a LUT for tone mapping
We would like a workflow for creating assets in Blender that closely
match the results when rendered in Hubs. Part of that is matching the
color grading. The best way to do this is to actually apply the exact
same color grading as is applied in blender, via a LUT exported from the
same OpenColorIO profiles.

This may also prove generally useful to scene creators wishing to use a
custom LUT for artistic reasons. This may make sense to try and land upstream.