[HUBS] Don't apply IBL irradiance to lightmapped objects
Based on our art pipeline we assume this information will already be
baked into lightmaps. This prevents applying this indirect lighting an
additional time, leading to objects being brighter than they should be.
A more complete implementation would let the user configure what
lights (and indirect light sources) should be considered "baked in" to a
lightmap. This would be part of a larger "light linking" strategy, but
is more than we need right now.